# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shop API
  version: 1.0.0
  description: |
    REST API магазина бытовой техники для управления клиентами, товарами,
    поставщиками и изображениями. Все идентификаторы имеют формат UUID.

    Все методы `/api/v1`, кроме регистрации, входа (в том числе через
    OAuth-провайдера) и сброса пароля, требуют access-токен в заголовке
    `Authorization: Bearer <token>`. Токен выдают методы
    `POST /api/v1/auth/register`, `POST /api/v1/auth/login` и
    `POST /api/v1/auth/oauth/callback`, а проверяет его сервис авторизации.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: Локальный сервер разработки
    variables:
      port:
        default: '8080'
        description: Порт HTTP-сервера приложения
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/products/{productId}/image:
    post:
      tags: [Images]
      summary: Добавить изображение товара
      description: Загружает бинарное изображение и связывает его с товаром.
      operationId: createProductImage
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Изображение добавлено
          headers:
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
              example:
                id: f67b8f8a-92d5-4de3-b991-b0fa45768bab
                product_id: 45fd8e62-2ce9-47d9-9616-fca5557bfb42
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PRODUCT_NOT_FOUND
                message: product not found
        '409':
          description: У товара уже есть изображение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: IMAGE_ALREADY_EXISTS
                message: product already has an image
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Images]
      summary: Получить изображение товара
      description: Возвращает изображение как загружаемый бинарный файл.
      operationId: getImageByProductId
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Бинарное содержимое изображения
          headers:
            Content-Disposition:
              $ref: '#/components/headers/ContentDisposition'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен из ответа регистрации или входа
  parameters:
...
    ProductId:
      name: productId
      in: path
      description: Идентификатор товара
      required: true
      schema:
        type: string
        format: uuid
      example: 45fd8e62-2ce9-47d9-9616-fca5557bfb42
...
  headers:
    Location:
      description: URI созданного ресурса
      schema:
        type: string
        format: uri-reference
      example: /api/v1/resources/8fd941ad-4a31-4d60-a379-cd4b13603a9a
    ContentDisposition:
      description: Указывает браузеру скачать ответ как файл
      schema:
        type: string
      example: 'attachment; filename="image.bin"'
  responses:
    Unauthorized:
      description: Access-токен отсутствует или недействителен
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: missing or invalid access token
    BadRequest:
      description: Некорректный запрос или ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: validation error
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: resource not found
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: internal server error
  schemas:
...
    ImageMetadata:
      type: object
      additionalProperties: false
      required: [id, product_id]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: f67b8f8a-92d5-4de3-b991-b0fa45768bab
        product_id:
          type: string
          format: uuid
          readOnly: true
          example: 45fd8e62-2ce9-47d9-9616-fca5557bfb42
...
    Error:
      type: object
      description: Код ошибки для обработки на фронтенде и человекочитаемое сообщение
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: VALIDATION_ERROR
        message:
          type: string
          description: Человекочитаемое сообщение об ошибке
          example: validation error